+++ app/src/main/java/net/appitiza/moderno/model/UserListdata.kt
@@ -5,6 +5,6 @@ import java.io.Serializable
 open class UserListdata : Serializable{
      var emailId: String? = null
      var username: String? = null
-     var salary: Int = 0
+     var salary: Double = 0.0
 
 }
